+++ src/lib/pages/migrate/components/MigrateContract.tsx
@@ -27,8 +27,8 @@ import {
 import { CodeSelectSection } from "lib/components/select-code";
 import { useSchemaStore } from "lib/providers/store";
 import { useTxBroadcast } from "lib/providers/tx-broadcast";
-import type { CodeIdInfoResponse } from "lib/services/code";
-import { useLCDCodeInfo } from "lib/services/codeService";
+import type { CodeInfoResponseLcd } from "lib/services/wasm/code";
+import { useCodeInfoLcd } from "lib/services/wasm/code";
 import type { BechAddr32, ComposedMsg } from "lib/types";
 import { MsgType } from "lib/types";
 import { composeMsg, isId, jsonValidate, resolvePermission } from "lib/utils";
@@ -120,13 +120,13 @@ export const MigrateContract = ({
     },
   });
 
-  const { refetch } = useLCDCodeInfo(codeId, {
+  const { refetch } = useCodeInfoLcd(codeId, {
     enabled: false,
     retry: false,
     cacheTime: 0,
     onSuccess(data) {
-      const permission = data.code_info.instantiate_permission;
-      setValue("codeHash", data.code_info.data_hash.toLowerCase());
+      const permission = data.codeInfo.instantiatePermission;
+      setValue("codeHash", data.codeInfo.dataHash.toLowerCase());
       if (
         resolvePermission(
           address,
@@ -256,8 +256,8 @@ export const MigrateContract = ({
           setValue("codeId", code);
           resetMsgInputSchema();
         }}
-        setCodeHash={(data: CodeIdInfoResponse) => {
-          setValue("codeHash", data.code_info.data_hash.toLowerCase());
+        setCodeHash={(data: CodeInfoResponseLcd) => {
+          setValue("codeHash", data.codeInfo.dataHash.toLowerCase());
         }}
         codeId={codeId}
       />